requirements.txt
mokuro>=0.2,<0.3
fastapi
uvicorn[standard]
pillow
